requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.1
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.6
filelock==3.16.1
fsspec==2024.10.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpie==3.2.3
httpx==0.27.2
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.11
networkx==3.4.2
openai==1.50.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.1
shellingham==1.5.4
smart-open==7.0.5
sniffio==1.3.1
sqlparse==0.5.1
srsly==2.5.0
threadpoolctl==3.5.0
tqdm==4.66.5
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
